config: remove the config handler in TiDB (#16022)
